+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 1c9dc78 - v11
+| Author: fjrodafo - Fri, 6 Dec 2024 14:13:36 +0000
+| 
 * Commit: 50959a7 - v10
 | Author: FJrodafo - Sun, 6 Oct 2024 19:35:14 +0100
 |